// src/app/models/calculos.model.ts
export interface Calculossoma{
    a: number; 
    b: number; 
    resultado: string; 
    paridade: string; 
    data: Date;
}

export interface Calculossubtracao{
    a: number; 
    b: number; 
    resultado: string; 
    paridade: string; 
    data: Date;
}

export interface Calculosdivisao{
    a: number; 
    b: number; 
    resultado: string; 
    paridade: string; 
    data: Date;
}

export interface Calculosmulti{
    a: number; 
    b: number; 
    resultado: string; 
    paridade: string; 
    data: Date;
}